pyproject.toml
[project]
name = "mcp-umls"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
